+++ src/gcd.h
@@ -1,7 +1,7 @@
 #ifndef __GCD__
 #define __GCD__
 
-inline int _gcd_recursive(unsigned int a, unsigned int b) {
+inline long long int _gcd_recursive(unsigned int a, unsigned int b) {
 
   if (b == 0) {
     return a;
